Widened existing EventCounter tests to include PerformanceCounters
